#18 Implement Layout Background Sets infrastructure

+++ shelton-htpc/SheltonHTPC.Configurator/Utils/DataTools.cs
@@ -2,6 +2,7 @@
 using SheltonHTPC.Data.Entities;
 using System;
 using System.IO;
+using System.Threading;
 using System.Threading.Tasks;
 using System.Windows;
 
@@ -11,9 +12,26 @@ namespace SheltonHTPC.Utils
     {
         public static async Task PurgeData(GeneralSettings settings)
         {
+            if (settings is null)
+                throw new ArgumentNullException(nameof(settings));
+
+            string dataPath = settings.DataPath;
+
             await Task.Run(() =>
             {
-                Directory.Delete(settings.DataPath, true);
+                int retries = 0;
+                while (Directory.Exists(dataPath) && retries < 5)
+                {
+                    try
+                    {
+                        Directory.Delete(dataPath, true);
+                    }
+                    catch (IOException)
+                    {
+                        Thread.Sleep(5000);
+                        ++retries;
+                    }
+                }
 
                 if (File.Exists(DataHelper.DataPathFilePath))
                     File.Delete(DataHelper.DataPathFilePath);
